Added styles to be able to accomodate more languages
