+++ src/locales/en.json
@@ -1,5 +1,8 @@
 {
   "translation": {
+    "title": "Tracking COVID-19",
+    "description": "Help the Civil Protection Team to track potential COVID-19 infections in Iceland",
+    "continue": "Continue in English",
     "aboutCovidDescription": "Sound advice and reliable information can be found on the covid.is website.",
     "aboutCovidTitle": "Information on COVID-19",
     "announcements": "Announcements",
