+++ src/locales/es.json
@@ -1,5 +1,8 @@
 {
   "translation": {
+    "title": "Rastreo COVID-19",
+    "description": "Ayuda al ñsks sjpsj  ss sEquipo de Protección Civil a rastrear posibles infecciones de COVID-19 en Islandia",
+    "continue": "Continuar en español",
     "aboutCovidDescription": "Puedes encontrar consejos e información fiable en la web covid.is/es",
     "aboutCovidTitle": "Información sobre el COVID-19",
     "announcements": "Anuncios",
@@ -31,7 +34,7 @@
     "icelandicAnswerLabel": "Respuesta ante el COVID-19",
     "icelandicAnswerLink": "https://www.covid.is/es/la-respuesta-islandesa-covid-19-que-es",
     "incorrentPINMessage": "¡Vaya! Código erróneo. Inténtelo de nuevo, por favor.",
-    "isolationLabel": "¿En qué consiste el aislamiento?",
+    "isolationLabel": "¿En qué puede consistir el aislamiento?",
     "isolationLink": "https://www.covid.is/es/contagio-confirmado-significa-aislamiento",
     "next": "Siguiente",
     "phoneNr": "Nº de teléfono",
