+++ src/locales/is.json
@@ -1,5 +1,8 @@
 {
   "translation": {
+    "title": "Rakning COVID-19",
+    "description": "Hjálpum rakningateymi Almannavarna að rekja hugsanleg COVID-19 smit á Íslandi",
+    "continue": "Áfram á íslensku",
     "aboutCovidDescription": "Á heimasíðunni covid.is er að finna ýmis góð ráð og traustar upplýsingar.",
     "aboutCovidTitle": "Upplýsingar um COVID-19",
     "announcements": "Tilkynningar",
