+++ src/locales/pl.json
@@ -1,5 +1,8 @@
 {
   "translation": {
+    "title": "Śledzenie COVID-19",
+    "description": "Pomoc Obronie Cywilnej w śledzeniu potencjalnych infekcji COVID-19 na Islandii",
+    "continue": "Prześlij do polskiego",
     "aboutCovidDescription": "Na stronie covid.is znajdziesz różne wskazówki i wiarygodne informacje.",
     "aboutCovidTitle": "Informacje o COVID-19",
     "announcements": "Ogłoszenia",
