--- src/router/logged-out/screens/LanguageScreen.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import LanguagePicker from '../../../components/LanguagePicker';
-import AppShell from '../../../components/AppShell';
-
-const LanguageScreen = () => (
-  <AppShell>
-    <LanguagePicker />
-  </AppShell>
-);
-
-export default LanguageScreen;
